LayoutTests/platform/mac-wk2/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/type-change-state-expected.txt


FAIL change state from hidden to text Type error
FAIL change state from hidden to search Type error
FAIL change state from hidden to tel Type error
FAIL change state from hidden to url Type error
FAIL change state from hidden to email Type error
FAIL change state from hidden to password Type error
FAIL change state from hidden to datetime-local Type error
FAIL change state from hidden to date Type error
FAIL change state from hidden to month Type error
FAIL change state from hidden to week Type error
FAIL change state from hidden to time Type error
FAIL change state from hidden to number Type error
FAIL change state from hidden to range Type error
FAIL change state from hidden to color Type error
FAIL change state from hidden to checkbox Type error
FAIL change state from hidden to radio Type error
PASS change state from hidden to file
FAIL change state from hidden to submit Type error
FAIL change state from hidden to image Type error
FAIL change state from hidden to reset Type error
FAIL change state from hidden to button Type error
FAIL change state from text to hidden Type error
PASS change state from text to search
PASS change state from text to tel
FAIL change state from text to url assert_equals: selectionStart should be unchanged expected 2 but got 6
FAIL change state from text to email assert_equals: selectionStart should be unchanged expected 2 but got 6
PASS change state from text to password
FAIL change state from text to datetime-local Type error
FAIL change state from text to date Type error
FAIL change state from text to month Type error
FAIL change state from text to week Type error
FAIL change state from text to time Type error
FAIL change state from text to number assert_equals: selectionStart should be unchanged expected 2 but got 0
FAIL change state from text to range Type error
FAIL change state from text to color Type error
FAIL change state from text to checkbox Type error
FAIL change state from text to radio Type error
PASS change state from text to file
FAIL change state from text to submit Type error
FAIL change state from text to image Type error
FAIL change state from text to reset Type error
FAIL change state from text to button Type error
FAIL change state from search to hidden Type error
PASS change state from search to text
PASS change state from search to tel
FAIL change state from search to url assert_equals: selectionStart should be unchanged expected 2 but got 6
FAIL change state from search to email assert_equals: selectionStart should be unchanged expected 2 but got 6
PASS change state from search to password
FAIL change state from search to datetime-local Type error
FAIL change state from search to date Type error
FAIL change state from search to month Type error
FAIL change state from search to week Type error
FAIL change state from search to time Type error
FAIL change state from search to number assert_equals: selectionStart should be unchanged expected 2 but got 0
FAIL change state from search to range Type error
FAIL change state from search to color Type error
FAIL change state from search to checkbox Type error
FAIL change state from search to radio Type error
PASS change state from search to file
FAIL change state from search to submit Type error
FAIL change state from search to image Type error
FAIL change state from search to reset Type error
FAIL change state from search to button Type error
FAIL change state from tel to hidden Type error
PASS change state from tel to text
PASS change state from tel to search
FAIL change state from tel to url assert_equals: selectionStart should be unchanged expected 2 but got 6
FAIL change state from tel to email assert_equals: selectionStart should be unchanged expected 2 but got 6
PASS change state from tel to password
FAIL change state from tel to datetime-local Type error
FAIL change state from tel to date Type error
FAIL change state from tel to month Type error
FAIL change state from tel to week Type error
FAIL change state from tel to time Type error
FAIL change state from tel to number assert_equals: selectionStart should be unchanged expected 2 but got 0
FAIL change state from tel to range Type error
FAIL change state from tel to color Type error
FAIL change state from tel to checkbox Type error
FAIL change state from tel to radio Type error
PASS change state from tel to file
FAIL change state from tel to submit Type error
FAIL change state from tel to image Type error
FAIL change state from tel to reset Type error
FAIL change state from tel to button Type error
FAIL change state from url to hidden Type error
PASS change state from url to text
PASS change state from url to search
PASS change state from url to tel
PASS change state from url to email
PASS change state from url to password
FAIL change state from url to datetime-local Type error
FAIL change state from url to date Type error
FAIL change state from url to month Type error
FAIL change state from url to week Type error
FAIL change state from url to time Type error
FAIL change state from url to number assert_equals: selectionStart should be unchanged expected 2 but got 0
FAIL change state from url to range Type error
FAIL change state from url to color Type error
FAIL change state from url to checkbox Type error
FAIL change state from url to radio Type error
PASS change state from url to file
FAIL change state from url to submit Type error
FAIL change state from url to image Type error
FAIL change state from url to reset Type error
FAIL change state from url to button Type error
FAIL change state from email to hidden Type error
PASS change state from email to text
PASS change state from email to search
PASS change state from email to tel
PASS change state from email to url
PASS change state from email to password
FAIL change state from email to datetime-local Type error
FAIL change state from email to date Type error
FAIL change state from email to month Type error
FAIL change state from email to week Type error
FAIL change state from email to time Type error
FAIL change state from email to number assert_equals: selectionStart should be unchanged expected 2 but got 0
FAIL change state from email to range Type error
FAIL change state from email to color Type error
FAIL change state from email to checkbox Type error
FAIL change state from email to radio Type error
PASS change state from email to file
FAIL change state from email to submit Type error
FAIL change state from email to image Type error
FAIL change state from email to reset Type error
FAIL change state from email to button Type error
FAIL change state from password to hidden Type error
PASS change state from password to text
PASS change state from password to search
PASS change state from password to tel
FAIL change state from password to url assert_equals: selectionStart should be unchanged expected 2 but got 6
FAIL change state from password to email assert_equals: selectionStart should be unchanged expected 2 but got 6
FAIL change state from password to datetime-local Type error
FAIL change state from password to date Type error
FAIL change state from password to month Type error
FAIL change state from password to week Type error
FAIL change state from password to time Type error
FAIL change state from password to number assert_equals: selectionStart should be unchanged expected 2 but got 0
FAIL change state from password to range Type error
FAIL change state from password to color Type error
FAIL change state from password to checkbox Type error
FAIL change state from password to radio Type error
PASS change state from password to file
FAIL change state from password to submit Type error
FAIL change state from password to image Type error
FAIL change state from password to reset Type error
FAIL change state from password to button Type error
FAIL change state from datetime-local to hidden Type error
FAIL change state from datetime-local to text Type error
FAIL change state from datetime-local to search Type error
FAIL change state from datetime-local to tel Type error
FAIL change state from datetime-local to url Type error
FAIL change state from datetime-local to email Type error
FAIL change state from datetime-local to password Type error
FAIL change state from datetime-local to date Type error
FAIL change state from datetime-local to month Type error
FAIL change state from datetime-local to week Type error
FAIL change state from datetime-local to time Type error
FAIL change state from datetime-local to number Type error
FAIL change state from datetime-local to range Type error
FAIL change state from datetime-local to color Type error
FAIL change state from datetime-local to checkbox Type error
FAIL change state from datetime-local to radio Type error
PASS change state from datetime-local to file
FAIL change state from datetime-local to submit Type error
FAIL change state from datetime-local to image Type error
FAIL change state from datetime-local to reset Type error
FAIL change state from datetime-local to button Type error
FAIL change state from date to hidden Type error
FAIL change state from date to text Type error
FAIL change state from date to search Type error
FAIL change state from date to tel Type error
FAIL change state from date to url Type error
FAIL change state from date to email Type error
FAIL change state from date to password Type error
FAIL change state from date to datetime-local Type error
FAIL change state from date to month Type error
FAIL change state from date to week Type error
FAIL change state from date to time Type error
FAIL change state from date to number Type error
FAIL change state from date to range Type error
FAIL change state from date to color Type error
FAIL change state from date to checkbox Type error
FAIL change state from date to radio Type error
PASS change state from date to file
FAIL change state from date to submit Type error
FAIL change state from date to image Type error
FAIL change state from date to reset Type error
FAIL change state from date to button Type error
FAIL change state from month to hidden Type error
FAIL change state from month to text Type error
FAIL change state from month to search Type error
FAIL change state from month to tel Type error
FAIL change state from month to url Type error
FAIL change state from month to email Type error
FAIL change state from month to password Type error
FAIL change state from month to datetime-local Type error
FAIL change state from month to date Type error
FAIL change state from month to week Type error
FAIL change state from month to time Type error
FAIL change state from month to number Type error
FAIL change state from month to range Type error
FAIL change state from month to color Type error
FAIL change state from month to checkbox Type error
FAIL change state from month to radio Type error
PASS change state from month to file
FAIL change state from month to submit Type error
FAIL change state from month to image Type error
FAIL change state from month to reset Type error
FAIL change state from month to button Type error
FAIL change state from week to hidden Type error
FAIL change state from week to text Type error
FAIL change state from week to search Type error
FAIL change state from week to tel Type error
FAIL change state from week to url Type error
FAIL change state from week to email Type error
FAIL change state from week to password Type error
FAIL change state from week to datetime-local Type error
FAIL change state from week to date Type error
FAIL change state from week to month Type error
FAIL change state from week to time Type error
FAIL change state from week to number Type error
FAIL change state from week to range Type error
FAIL change state from week to color Type error
FAIL change state from week to checkbox Type error
FAIL change state from week to radio Type error
PASS change state from week to file
FAIL change state from week to submit Type error
FAIL change state from week to image Type error
FAIL change state from week to reset Type error
FAIL change state from week to button Type error
FAIL change state from time to hidden Type error
FAIL change state from time to text Type error
FAIL change state from time to search Type error
FAIL change state from time to tel Type error
FAIL change state from time to url Type error
FAIL change state from time to email Type error
FAIL change state from time to password Type error
FAIL change state from time to datetime-local Type error
FAIL change state from time to date Type error
FAIL change state from time to month Type error
FAIL change state from time to week Type error
FAIL change state from time to number Type error
FAIL change state from time to range Type error
FAIL change state from time to color Type error
FAIL change state from time to checkbox Type error
FAIL change state from time to radio Type error
PASS change state from time to file
FAIL change state from time to submit Type error
FAIL change state from time to image Type error
FAIL change state from time to reset Type error
FAIL change state from time to button Type error
FAIL change state from number to hidden Type error
PASS change state from number to text
PASS change state from number to search
PASS change state from number to tel
PASS change state from number to url
PASS change state from number to email
PASS change state from number to password
FAIL change state from number to datetime-local Type error
FAIL change state from number to date Type error
FAIL change state from number to month Type error
FAIL change state from number to week Type error
FAIL change state from number to time Type error
FAIL change state from number to range Type error
FAIL change state from number to color Type error
FAIL change state from number to checkbox assert_equals: input.value should be 'on' after change of state expected "on" but got ""
FAIL change state from number to radio assert_equals: input.value should be 'on' after change of state expected "on" but got ""
PASS change state from number to file
FAIL change state from number to submit Type error
FAIL change state from number to image Type error
FAIL change state from number to reset Type error
FAIL change state from number to button Type error
FAIL change state from range to hidden Type error
FAIL change state from range to text Type error
FAIL change state from range to search Type error
FAIL change state from range to tel Type error
FAIL change state from range to url Type error
FAIL change state from range to email Type error
FAIL change state from range to password Type error
FAIL change state from range to datetime-local Type error
FAIL change state from range to date Type error
FAIL change state from range to month Type error
FAIL change state from range to week Type error
FAIL change state from range to time Type error
FAIL change state from range to number Type error
FAIL change state from range to color Type error
FAIL change state from range to checkbox Type error
FAIL change state from range to radio Type error
PASS change state from range to file
FAIL change state from range to submit Type error
FAIL change state from range to image Type error
FAIL change state from range to reset Type error
FAIL change state from range to button Type error
FAIL change state from color to hidden Type error
FAIL change state from color to text Type error
FAIL change state from color to search Type error
FAIL change state from color to tel Type error
FAIL change state from color to url Type error
FAIL change state from color to email Type error
FAIL change state from color to password Type error
FAIL change state from color to datetime-local Type error
FAIL change state from color to date Type error
FAIL change state from color to month Type error
FAIL change state from color to week Type error
FAIL change state from color to time Type error
FAIL change state from color to number Type error
FAIL change state from color to range Type error
FAIL change state from color to checkbox Type error
FAIL change state from color to radio Type error
PASS change state from color to file
FAIL change state from color to submit Type error
FAIL change state from color to image Type error
FAIL change state from color to reset Type error
FAIL change state from color to button Type error
FAIL change state from checkbox to hidden Type error
FAIL change state from checkbox to text Type error
FAIL change state from checkbox to search Type error
FAIL change state from checkbox to tel Type error
FAIL change state from checkbox to url Type error
FAIL change state from checkbox to email Type error
FAIL change state from checkbox to password Type error
FAIL change state from checkbox to datetime-local Type error
FAIL change state from checkbox to date Type error
FAIL change state from checkbox to month Type error
FAIL change state from checkbox to week Type error
FAIL change state from checkbox to time Type error
FAIL change state from checkbox to number Type error
FAIL change state from checkbox to range Type error
FAIL change state from checkbox to color Type error
FAIL change state from checkbox to radio Type error
PASS change state from checkbox to file
FAIL change state from checkbox to submit Type error
FAIL change state from checkbox to image Type error
FAIL change state from checkbox to reset Type error
FAIL change state from checkbox to button Type error
FAIL change state from radio to hidden Type error
FAIL change state from radio to text Type error
FAIL change state from radio to search Type error
FAIL change state from radio to tel Type error
FAIL change state from radio to url Type error
FAIL change state from radio to email Type error
FAIL change state from radio to password Type error
FAIL change state from radio to datetime-local Type error
FAIL change state from radio to date Type error
FAIL change state from radio to month Type error
FAIL change state from radio to week Type error
FAIL change state from radio to time Type error
FAIL change state from radio to number Type error
FAIL change state from radio to range Type error
FAIL change state from radio to color Type error
FAIL change state from radio to checkbox Type error
PASS change state from radio to file
FAIL change state from radio to submit Type error
FAIL change state from radio to image Type error
FAIL change state from radio to reset Type error
FAIL change state from radio to button Type error
PASS change state from file to hidden
PASS change state from file to text
PASS change state from file to search
PASS change state from file to tel
PASS change state from file to url
PASS change state from file to email
PASS change state from file to password
PASS change state from file to datetime-local
PASS change state from file to date
PASS change state from file to month
PASS change state from file to week
PASS change state from file to time
PASS change state from file to number
PASS change state from file to range
PASS change state from file to color
PASS change state from file to checkbox
PASS change state from file to radio
PASS change state from file to submit
PASS change state from file to image
PASS change state from file to reset
PASS change state from file to button
FAIL change state from submit to hidden Type error
FAIL change state from submit to text Type error
FAIL change state from submit to search Type error
FAIL change state from submit to tel Type error
FAIL change state from submit to url Type error
FAIL change state from submit to email Type error
FAIL change state from submit to password Type error
FAIL change state from submit to datetime-local Type error
FAIL change state from submit to date Type error
FAIL change state from submit to month Type error
FAIL change state from submit to week Type error
FAIL change state from submit to time Type error
FAIL change state from submit to number Type error
FAIL change state from submit to range Type error
FAIL change state from submit to color Type error
FAIL change state from submit to checkbox Type error
FAIL change state from submit to radio Type error
PASS change state from submit to file
FAIL change state from submit to image Type error
FAIL change state from submit to reset Type error
FAIL change state from submit to button Type error
FAIL change state from image to hidden Type error
FAIL change state from image to text Type error
FAIL change state from image to search Type error
FAIL change state from image to tel Type error
FAIL change state from image to url Type error
FAIL change state from image to email Type error
FAIL change state from image to password Type error
FAIL change state from image to datetime-local Type error
FAIL change state from image to date Type error
FAIL change state from image to month Type error
FAIL change state from image to week Type error
FAIL change state from image to time Type error
FAIL change state from image to number Type error
FAIL change state from image to range Type error
FAIL change state from image to color Type error
FAIL change state from image to checkbox Type error
FAIL change state from image to radio Type error
PASS change state from image to file
FAIL change state from image to submit Type error
FAIL change state from image to reset Type error
FAIL change state from image to button Type error
FAIL change state from reset to hidden Type error
FAIL change state from reset to text Type error
FAIL change state from reset to search Type error
FAIL change state from reset to tel Type error
FAIL change state from reset to url Type error
FAIL change state from reset to email Type error
FAIL change state from reset to password Type error
FAIL change state from reset to datetime-local Type error
FAIL change state from reset to date Type error
FAIL change state from reset to month Type error
FAIL change state from reset to week Type error
FAIL change state from reset to time Type error
FAIL change state from reset to number Type error
FAIL change state from reset to range Type error
FAIL change state from reset to color Type error
FAIL change state from reset to checkbox Type error
FAIL change state from reset to radio Type error
PASS change state from reset to file
FAIL change state from reset to submit Type error
FAIL change state from reset to image Type error
FAIL change state from reset to button Type error
FAIL change state from button to hidden Type error
FAIL change state from button to text Type error
FAIL change state from button to search Type error
FAIL change state from button to tel Type error
FAIL change state from button to url Type error
FAIL change state from button to email Type error
FAIL change state from button to password Type error
FAIL change state from button to datetime-local Type error
FAIL change state from button to date Type error
FAIL change state from button to month Type error
FAIL change state from button to week Type error
FAIL change state from button to time Type error
FAIL change state from button to number Type error
FAIL change state from button to range Type error
FAIL change state from button to color Type error
FAIL change state from button to checkbox Type error
FAIL change state from button to radio Type error
PASS change state from button to file
FAIL change state from button to submit Type error
FAIL change state from button to image Type error
FAIL change state from button to reset Type error
